Updated DB Schema
DB is now normalized so recipes can have arbitrary numbers of ingredients and steps. The RecipeDbAdapter can load and store these values into the Recipe object for convenience.

+++ androidapp/src/com/CookingAssistant/RecipeDbAdapter.java
@@ -1,20 +1,25 @@
 package com.CookingAssistant;
 
+import java.util.ArrayList;
+
 import android.content.ContentValues;
 import android.content.Context;
 import android.database.Cursor;
 import android.database.SQLException;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
+import android.database.sqlite.SQLiteQueryBuilder;
 import android.util.Log;
 
 
 public class RecipeDbAdapter {
 
 	public static final String KEY_TITLE = "title";
-    public static final String KEY_INGREDIENTS = "ingredients";
-    public static final String KEY_INSTRUCTIONS = "instructions";
+	public static final String KEY_FAVORITE = "favorite";
+    public static final String KEY_INGREDIENT = "ingredient";
+    public static final String KEY_STEP = "instruction";
     public static final String KEY_ROWID = "_id";
+    public static final String KEY_RECIPEID = "recipe";
 
     private static final String TAG = "RecipeDbAdapter";
     private DatabaseHelper mDbHelper;
@@ -23,16 +28,28 @@ public class RecipeDbAdapter {
     /**
      * Database creation sql statement
      */
-    private static final String DATABASE_CREATE =
-            "create table recipes ("
-    		+ "_id integer primary key autoincrement, "
-            + "title text not null, "
-            + "ingredients text not null, "
-            + "instructions text not null);";
+    private static final String DATABASE_CREATE_RECIPE_TABLE =
+            "create table recipe ("
+    			+ "_id integer primary key autoincrement, "
+    			+ "title text not null, "
+    			+ "favorite integer); ";
+    private static final String DATABASE_CREATE_INGREDIENT_TABLE =
+    		"create table ingredient ("
+    			+ "_id integer primary key autoincrement, "
+    			+ "foreign key (recipe) references recipe(_id), "
+    			+ "ingredient text not null); ";
+    private static final String DATABASE_CREATE_STEP_TABLE = 
+            "create table step ("
+            	+ "_id integer primary key autoincrement, "
+            	+ "foreign key (recipe) references recipe(_id), "
+            	+ "step text not null); ";
 
     private static final String DATABASE_NAME = "data";
-    public static final String DATABASE_TABLE = "recipes";
-    private static final int DATABASE_VERSION = 5;
+    private static final String DATABASE_RECIPE_TABLE = "recipes";
+    private static final String DATABASE_INGREDIENT_TABLE = "ingredient";
+    private static final String DATABASE_STEP_TABLE = "step";
+    
+    private static final int DATABASE_VERSION = 6;
 
     private final Context mCtx;
 
@@ -44,34 +61,17 @@ public class RecipeDbAdapter {
 
         @Override
         public void onCreate(SQLiteDatabase db) {
-            db.execSQL(DATABASE_CREATE);
-            db.execSQL("INSERT INTO "
-                    + DATABASE_TABLE
-                    + " (title, ingredients, instructions)"
-                    + " VALUES ('First Recipe', '1 sugar\n2 salt\nwater', 'mix sugar with salt\nadd water');");
-            db.execSQL("INSERT INTO "
-                    + DATABASE_TABLE
-                    + " (title, ingredients, instructions)"
-                    + " VALUES ('Second Recipe', 'water noodles', 'boil water add noodles/neat');");
-            db.execSQL("INSERT INTO "
-                    + DATABASE_TABLE
-                    + " (title, ingredients, instructions)"
-                    + " VALUES ('Third Recipe', 'two licorice', 'eat and chew a lot');");
-            db.execSQL("INSERT INTO "
-                    + DATABASE_TABLE
-                    + " (title, ingredients, instructions)"
-                    + " VALUES ('Fourth Recipe', 'steak', 'buy steak');");
-            db.execSQL("INSERT INTO "
-                    + DATABASE_TABLE
-                    + " (title, ingredients, instructions)"
-                    + " VALUES ('Fifth Recipe', 'skittles', 'taste the rainbow');");
+            db.execSQL(DATABASE_CREATE_RECIPE_TABLE);
+            db.execSQL(DATABASE_CREATE_INGREDIENT_TABLE);
+            db.execSQL(DATABASE_CREATE_STEP_TABLE);
         }
 
         @Override
         public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
             Log.w(TAG, "Upgrading database from version " + oldVersion + " to "
                     + newVersion + ", which will destroy all old data");
-            db.execSQL("DROP TABLE IF EXISTS recipes");
+            if (oldVersion < 6)
+            	db.execSQL("DROP TABLE IF EXISTS recipes");
             onCreate(db);
         }
     }
@@ -107,7 +107,7 @@ public class RecipeDbAdapter {
 
 
     /**
-     * Create a new recipe using the title and body provided. If the recipe is
+     * Create a new recipe using the content from the given Recipe. If the recipe is
      * successfully created return the new rowId for that recipe, otherwise return
      * a -1 to indicate failure.
      * 
@@ -115,13 +115,27 @@ public class RecipeDbAdapter {
      * @param body the body of the recipe
      * @return rowId or -1 if failed
      */
-    public long createRecipe(String title, String ingredients, String instructions) {
-        ContentValues initialValues = new ContentValues();
-        initialValues.put(KEY_TITLE, title);
-        initialValues.put(KEY_INGREDIENTS, ingredients);
-        initialValues.put(KEY_INSTRUCTIONS, instructions);
+    public long createRecipe(Recipe recipe) {
+        ContentValues recipeValues = new ContentValues();
+        recipeValues.put(KEY_TITLE, recipe.title);
+        recipeValues.put(KEY_FAVORITE, recipe.favorite);
+        long recipeId = mDb.insertOrThrow(DATABASE_RECIPE_TABLE, null, recipeValues);
 
-        return mDb.insert(DATABASE_TABLE, null, initialValues);
+    	recipeValues.clear();
+        for (String ingredient : recipe.ingredients) {
+        	recipeValues.put(KEY_RECIPEID, recipeId);
+        	recipeValues.put(KEY_INGREDIENT, ingredient);
+        	mDb.insert(DATABASE_INGREDIENT_TABLE, null, recipeValues);
+        }
+        
+    	recipeValues.clear();
+        for (String step : recipe.steps) {
+        	recipeValues.put(KEY_RECIPEID, recipeId);
+        	recipeValues.put(KEY_STEP, step);
+        	mDb.insert(DATABASE_STEP_TABLE, null, recipeValues);
+        }
+
+        return recipeId;
     }
 
     /**
@@ -131,8 +145,12 @@ public class RecipeDbAdapter {
      * @return true if deleted, false otherwise
      */
     public boolean deleteRecipe(long rowId) {
-
-        return mDb.delete(DATABASE_TABLE, KEY_ROWID + "=" + rowId, null) > 0;
+    	boolean deleted = mDb.delete(DATABASE_RECIPE_TABLE, KEY_ROWID + "=" + rowId, null) > 0;
+    	if (deleted) {
+    		mDb.delete(DATABASE_INGREDIENT_TABLE, KEY_RECIPEID + "=" + rowId, null);
+    		mDb.delete(DATABASE_STEP_TABLE, KEY_RECIPEID + "=" + rowId, null);
+    	}
+    	return deleted;
     }
 
     /**
@@ -141,33 +159,75 @@ public class RecipeDbAdapter {
      * @return Cursor over all recipes
      */
     public Cursor fetchAllRecipes() {
-
-        return mDb.query(DATABASE_TABLE, new String[] {KEY_ROWID, KEY_TITLE,
-                KEY_INGREDIENTS, KEY_INSTRUCTIONS}, null, null, null, null, null);
+        return mDb.query(DATABASE_RECIPE_TABLE, new String[] {KEY_ROWID, KEY_TITLE, KEY_FAVORITE}, null, null, null, null, null);
     }
 
     /**
-     * Return a Cursor positioned at the recipe that matches the given rowId
+     * Return a Recipe loaded from the given rowId
      * 
      * @param rowId id of recipe to retrieve
-     * @return Cursor positioned to matching recipe, if found
+     * @return Recipe loaded with values for given recipe
      * @throws SQLException if recipe could not be found/retrieved
      */
-    public Cursor fetchRecipe(long rowId) throws SQLException {
-        Cursor mCursor =
-                mDb.query(true, DATABASE_TABLE, new String[] {KEY_ROWID,
-                        KEY_TITLE, KEY_INGREDIENTS, KEY_INSTRUCTIONS}, KEY_ROWID + "=" + rowId, null,
-                        null, null, null, null);
+    public Recipe fetchRecipe(long rowId) throws SQLException {
+    	Recipe recipe = new Recipe();
+    	
+    	// General recipe info
+        Cursor mCursor = 
+        	mDb.query(true, DATABASE_RECIPE_TABLE, 
+        			  new String[] {KEY_ROWID, KEY_TITLE, KEY_FAVORITE},
+        			  KEY_ROWID + "=" + rowId, null,
+                      null, null, null, null);
         if (mCursor != null) {
             mCursor.moveToFirst();
+            recipe.id = mCursor.getLong(mCursor.getColumnIndex(KEY_ROWID));
+            recipe.title = mCursor.getString(mCursor.getColumnIndex(KEY_TITLE));
+            recipe.favorite = mCursor.getInt(mCursor.getColumnIndex(KEY_FAVORITE)) != 0;
         }
-        return mCursor;
+        mCursor.close();
+        
+        // Ingredients
+        ArrayList<String> ingredients = new ArrayList<String> ();
+        mCursor =
+        	mDb.query(false, DATABASE_INGREDIENT_TABLE,
+        			  new String[] {KEY_INGREDIENT},
+        			  KEY_RECIPEID + "=" + rowId, null,
+        			  null, null, null, null);
+        if (mCursor != null) {
+        	mCursor.moveToFirst();
+            while (mCursor.isAfterLast() == false) {
+                ingredients.add(mCursor.getString(mCursor.getColumnIndex(KEY_INGREDIENT)));
+                mCursor.moveToNext();
+            }
+        }
+        recipe.ingredients = (String[]) ingredients.toArray();
+        mCursor.close();
+        
+        // Steps
+        ArrayList<String> steps = new ArrayList<String> ();
+        mCursor =
+        	mDb.query(false, DATABASE_STEP_TABLE,
+        			  new String[] {KEY_STEP},
+        			  KEY_RECIPEID + "=" + rowId, null,
+        			  null, null, null, null);
+        if (mCursor != null) {
+        	mCursor.moveToFirst();
+            while (mCursor.isAfterLast() == false) {
+                steps.add(mCursor.getString(mCursor.getColumnIndex(KEY_STEP)));
+                mCursor.moveToNext();
+            }
+        }
+        recipe.steps = (String[]) steps.toArray();
+        mCursor.close();
+        
+        return recipe;
     }
     
     public Cursor searchRecipes(String s) {
+    	// FIXME: Need to query against the ingredients and steps!
     	Cursor mCursor =
-            mDb.query(true, DATABASE_TABLE, new String[] {KEY_ROWID,
-                    KEY_TITLE, KEY_INGREDIENTS, KEY_INSTRUCTIONS}, KEY_TITLE + " LIKE " + '?', 
+            mDb.query(true, DATABASE_RECIPE_TABLE, new String[] {KEY_ROWID,
+                    KEY_TITLE, KEY_FAVORITE}, KEY_TITLE + " LIKE " + '?', 
                     new String[]{"%" + s + "%"},
                     null, null, null, null);
     	if (mCursor != null) {
@@ -177,7 +237,7 @@ public class RecipeDbAdapter {
     }
 
     /**
-     * Update the recipe using the details provided. The recipe to be updated is
+     * Update the recipe using the details provided in Recipe. The recipe to be updated is
      * specified using the rowId, and it is altered to use the title and body
      * values passed in
      * 
@@ -186,12 +246,14 @@ public class RecipeDbAdapter {
      * @param body value to set recipe body to
      * @return true if the recipe was successfully updated, false otherwise
      */
-    public boolean updaterecipe(long rowId, String title, String ingredients, String instructions) {
+    public boolean updaterecipe(Recipe recipe) {
         ContentValues args = new ContentValues();
-        args.put(KEY_TITLE, title);
-        args.put(KEY_INGREDIENTS, ingredients);
-        args.put(KEY_INSTRUCTIONS, instructions);
+        args.put(KEY_TITLE, recipe.title);
+        args.put(KEY_FAVORITE, recipe.favorite);
+        boolean result = mDb.update(DATABASE_RECIPE_TABLE, args, KEY_ROWID + "=" + recipe.id, null) > 0;
+        
+        // FIXME: Need to also update ingredients and steps
 
-        return mDb.update(DATABASE_TABLE, args, KEY_ROWID + "=" + rowId, null) > 0;
+        return result;
     }
 }
